fix(clippy): rename mod forward_index to index (#517)
