--- tskv/src/forward_index/mod.rs
@@ -1,6 +1,0 @@
-mod errors;
-mod forward_index;
-mod tests;
-
-pub use errors::*;
-pub use forward_index::*;
